Improve language in title check, never return with language
This also added an encoding library name check as well. This was changed to never return a title that has a Language found within it. It also now uses Language.find to try scrub the title string for a language by name (generally anywhere within the string) or tag (if generally whole string is a valid tag).

The efficiency of the alphanumeric codec check has also been improved for cases like H.264 and VC-1. Mono and Atmos channel checks have been added, and the channel layout (float) check has been removed as it is too common to be used in actual names. E.g. `Video Edit 2.0`.