Remove Language Tag check from Track.title due to false-flags
E.g., `SDH` as a title can be detected as Southern Kurdish, when it meant to mean `Subtitles for the Deaf or Hard-of-Hearing`, i.e., not a language.

The probability of a language tag being used as a title is very low anyway, and if that as a title looked bad in a template the caller could just change either the template or modify the input file.