No longer filter Track title in any way
The possibility for information loss outweighs the benefits, especially when the checks for the benefits are pretty hacky.

Now, templates simply have to assume the title is used for it's intended purpose, i.e., a unique track name with actual useful info not set anywhere else in the track metadata.