Create an advanced example template (MiU)
This template shows how to introspect and modify the incoming NFO data in multiple ways. E.g., if the template is for a description, it detects that and does some changes like center aligning the Art and NFO.

This also shows stuff like moving the release name, doing conditional placement, and using imports.